version .3

+++ database/migrations/2021_09_29_153631_create_bibliografia_digital_table.php
@@ -17,7 +17,7 @@ class CreateBibliografiaDigitalTable extends Migration
             $table->string('clasificacion',40)->unique();
             $table->string('titulo',60);
             $table->string('autor',50);
-            $table->bigInteger('anio');
+            $table->integer('anio');
             $table->softDeletes();
             $table->timestamps();
         });
